Replace `device::DeviceManager` with the new system
