+++ device-manager/src/main.rs
@@ -20,7 +20,6 @@ extern crate seshat;
 extern crate tags;
 
 // Local modules
-mod alt_device;
 mod device;
 mod indexer;
 mod logging;
@@ -59,8 +58,7 @@ fn main() {
 
     // Create the device manager
     let (tx, cancel) = mpsc::channel();
-    // let manager = device::DeviceManager::new(index, tx.clone());
-    let manager = alt_device::DeviceManager::new(index, tx.clone());
+    let manager = device::DeviceManager::new(index, tx.clone());
     trace!("Created device state manager");
 
     // TODO: This currently hangs until the indexing has completed
